refactor: move ErrorResponse type to shared/api folder
- Update import paths for ErrorResponse type
